+++ adapters/wagmi-connector/src/index.ts
@@ -1,4 +1,2 @@
-export { default as BloctoConnector } from './bloctoConnector';
-export { default as BloctoWeb3ModalConfig } from './bloctoWeb3ModalConfig';
-
-export type * from './bloctoConnector';
+export { default as BloctoConnector } from './connector';
+export type * from './connector';