+++ packages/blocto-sdk/src/index.d.ts
@@ -7,15 +7,10 @@ import {
   EthereumProviderInterface,
 } from './providers/types/ethereum.d';
 import {
-  SolanaProviderConfig,
-  SolanaProviderInterface,
-} from './providers/types/solana.d';
-import {
   AptosProviderConfig,
   AptosProviderInterface,
 } from './providers/types/aptos.d';
 import * as AptosTypes from './providers/types/aptos.d';
-import * as SolanaTypes from './providers/types/solana.d';
 import * as EthereumTypes from './providers/types/ethereum.d';
 
 export * from './providers/types/blocto.d';
@@ -23,23 +18,18 @@ export {
   BaseConfig,
   EthereumProviderConfig,
   EthereumProviderInterface,
-  SolanaProviderConfig,
-  SolanaProviderInterface,
   AptosProviderConfig,
   AptosProviderInterface,
   // Keep above types for backward compatibility
   AptosTypes,
-  SolanaTypes,
   EthereumTypes,
 };
 export declare interface BloctoSDKConfig extends BaseConfig {
   ethereum?: Omit<EthereumProviderConfig, 'appId'>;
-  solana?: Omit<SolanaProviderConfig, 'appId'>;
   aptos?: Omit<AptosProviderConfig>;
 }
 declare class BloctoSDK {
   ethereum?: EthereumProviderInterface;
-  solana?: SolanaProviderInterface;
   aptos?: AptosProviderInterface;
   constructor(config: BloctoSDKConfig);
 }