+++ packages/blocto-sdk/CHANGELOG.md
@@ -1,5 +1,78 @@
 # @blocto/sdk
 
+## 0.6.0
+
+### Minor Changes
+
+- e0e8fb7: Add switch chain hint ui for evm-provider
+- 34a23e4: DEPRECATED: Solana provider no longer supported and removed.
+
+### Patch Changes
+
+- 21097d9: Fix switch chain is connected logic
+- 6e427b1: handle potential undefined when passing unsupported chainId into bloctoSDK
+- 55f3395: Fix go login flow when switching to a different blocto server
+- 3ceb547: Fix skip switch chain if provide same id as current
+- 21097d9: Fix disconnect all evm chains when disconnect
+- 21097d9: Fix wrong switch chain url
+- 21097d9: Fix switch chain login logic
+- 21097d9: Fix emit disconnect event when switch chain approved but failed
+- 494ded1: enhance tx params check
+
+## 0.6.0-beta.7
+
+### Patch Changes
+
+- 6e427b1: handle potential undefined when passing unsupported chainId into bloctoSDK
+- 55f3395: Fix go login flow when switching to a different blocto server
+- 494ded1: enhance tx params check
+
+## 0.6.0-beta.6
+
+### Minor Changes
+
+- 34a23e4: DEPRECATED: Solana provider no longer supported and removed.
+
+### Patch Changes
+
+- 3ceb547: Fix skip switch chain if provide same id as current
+
+## 0.6.0-beta.5
+
+### Patch Changes
+
+- 39754e5: Fix disconnect all evm chains when disconnect
+
+## 0.6.0-beta.4
+
+### Patch Changes
+
+- 6acf18d: Fix switch chain is connected logic
+
+## 0.6.0-beta.3
+
+### Patch Changes
+
+- 84d6727: Fix emit disconnect event when switch chain approved but failed
+
+## 0.6.0-beta.2
+
+### Patch Changes
+
+- a0cc54c: Fix switch chain login logic
+
+## 0.6.0-beta.1
+
+### Patch Changes
+
+- 0d5bfd5: Fix wrong switch chain url
+
+## 0.6.0-beta.0
+
+### Minor Changes
+
+- e0e8fb7: Add switch chain hint ui for evm-provider
+
 ## 0.5.5
 
 ### Patch Changes